# labo-gestion-front/src/openapi/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenApi specification - Samaraza","description":"OpenApi documentation for Spring Security","termsOfService":"Terms of service","contact":{"name":"Samaraza","url":"https://samaraza.com","email":"samaraza@email.com"},"license":{"name":"Licence name","url":"https://some-url.com"},"version":"1.0"},"servers":[{"url":"http://localhost:8089/api/v1","description":"Local ENV"},{"url":"https://samaraza-labo-gestion.com/api/v1","description":"PROD ENV"}],"security":[{"bearerAuth":[]}],"paths":{"/users/{id}":{"get":{"tags":["user-controller"],"operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"put":{"tags":["user-controller"],"operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"delete":{"tags":["user-controller"],"operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/users/admin/{id}":{"put":{"tags":["user-controller"],"operationId":"updateFromAdmin","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/tps/{id}":{"get":{"tags":["tp-controller"],"operationId":"getTpById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Tp"}}}}}},"put":{"tags":["tp-controller"],"operationId":"updateTp","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tp"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Tp"}}}}}},"delete":{"tags":["tp-controller"],"operationId":"deleteTp","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/salle-tps/{id}":{"get":{"tags":["salle-tp-controller"],"operationId":"getSalleTpById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SalleTp"}}}}}},"put":{"tags":["salle-tp-controller"],"operationId":"updateSalleTp","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SalleTp"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SalleTp"}}}}}},"delete":{"tags":["salle-tp-controller"],"operationId":"deleteSalleTp","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/produits/{id}":{"get":{"tags":["produit-controller"],"operationId":"getProduitById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Produit"}}}}}},"put":{"tags":["produit-controller"],"operationId":"updateProduit","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Produit"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Produit"}}}}}},"delete":{"tags":["produit-controller"],"operationId":"deleteProduit","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/preparations/{id}":{"get":{"tags":["preparation-controller"],"operationId":"getPreparationById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Preparation"}}}}}},"put":{"tags":["preparation-controller"],"operationId":"updatePreparation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Preparation"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Preparation"}}}}}},"delete":{"tags":["preparation-controller"],"operationId":"deletePreparation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/posts/{id}":{"get":{"tags":["post-controller"],"operationId":"getPostById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Post"}}}}}},"put":{"tags":["post-controller"],"operationId":"updatePost","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Post"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Post"}}}}}},"delete":{"tags":["post-controller"],"operationId":"deletePost","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/labos/{id}":{"get":{"tags":["labo-controller"],"operationId":"getLaboById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Labo"}}}}}},"put":{"tags":["labo-controller"],"operationId":"updateLabo","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Labo"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Labo"}}}}}},"delete":{"tags":["labo-controller"],"operationId":"deleteLabo","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/inventaires/{id}":{"get":{"tags":["inventaire-controller"],"operationId":"getInventaireById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Inventaire"}}}}}},"put":{"tags":["inventaire-controller"],"operationId":"updateInventaire","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Inventaire"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Inventaire"}}}}}},"delete":{"tags":["inventaire-controller"],"operationId":"deleteInventaire","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/fournisseurs/{id}":{"get":{"tags":["fournisseur-controller"],"operationId":"getFournisseurById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Fournisseur"}}}}}},"put":{"tags":["fournisseur-controller"],"operationId":"updateFournisseur","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Fournisseur"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Fournisseur"}}}}}},"delete":{"tags":["fournisseur-controller"],"operationId":"deleteFournisseur","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/commandes/{id}":{"get":{"tags":["commande-controller"],"operationId":"getCommandeById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Commande"}}}}}},"put":{"tags":["commande-controller"],"operationId":"updateCommande","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Commande"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Commande"}}}}}},"delete":{"tags":["commande-controller"],"operationId":"deleteCommande","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/armoires/{id}":{"get":{"tags":["armoire-controller"],"operationId":"getArmoireById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Armoire"}}}}}},"put":{"tags":["armoire-controller"],"operationId":"updateArmoire","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Armoire"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Armoire"}}}}}},"delete":{"tags":["armoire-controller"],"operationId":"deleteArmoire","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/users":{"get":{"tags":["user-controller"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}},"post":{"tags":["user-controller"],"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/tps":{"get":{"tags":["tp-controller"],"operationId":"getAllTps","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Tp"}}}}}}},"post":{"tags":["tp-controller"],"operationId":"createTp","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tp"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Tp"}}}}}}},"/salle-tps":{"get":{"tags":["salle-tp-controller"],"operationId":"getAllSalleTps","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SalleTp"}}}}}}},"post":{"tags":["salle-tp-controller"],"operationId":"createSalleTp","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SalleTp"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SalleTp"}}}}}}},"/produits":{"get":{"tags":["produit-controller"],"operationId":"getAllProduits","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Produit"}}}}}}},"post":{"tags":["produit-controller"],"operationId":"createProduit","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Produit"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Produit"}}}}}}},"/preparations":{"get":{"tags":["preparation-controller"],"operationId":"getAllPreparations","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Preparation"}}}}}}},"post":{"tags":["preparation-controller"],"operationId":"createPreparation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Preparation"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Preparation"}}}}}}},"/posts":{"get":{"tags":["post-controller"],"operationId":"getAllPosts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Post"}}}}}}},"post":{"tags":["post-controller"],"operationId":"createPost","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Post"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Post"}}}}}}},"/labos":{"get":{"tags":["labo-controller"],"operationId":"getAllLabos","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Labo"}}}}}}},"post":{"tags":["labo-controller"],"operationId":"createLabo","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Labo"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Labo"}}}}}}},"/inventaires":{"get":{"tags":["inventaire-controller"],"operationId":"getAllInventaires","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Inventaire"}}}}}}},"post":{"tags":["inventaire-controller"],"operationId":"createInventaire","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Inventaire"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Inventaire"}}}}}}},"/fournisseurs":{"get":{"tags":["fournisseur-controller"],"operationId":"getAllFournisseurs","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Fournisseur"}}}}}}},"post":{"tags":["fournisseur-controller"],"operationId":"createFournisseur","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Fournisseur"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Fournisseur"}}}}}}},"/commandes":{"get":{"tags":["commande-controller"],"operationId":"getAllCommandes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Commande"}}}}}}},"post":{"tags":["commande-controller"],"operationId":"createCommande","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Commande"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Commande"}}}}}}},"/auth/register":{"post":{"tags":["Authentication"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationRequest"}}},"required":true},"responses":{"202":{"description":"Accepted","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/auth/authenticate":{"post":{"tags":["Authentication"],"operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/armoires":{"get":{"tags":["armoire-controller"],"operationId":"getAllArmoires","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Armoire"}}}}}}},"post":{"tags":["armoire-controller"],"operationId":"createArmoire","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Armoire"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Armoire"}}}}}}},"/users/email/{email}":{"get":{"tags":["user-controller"],"operationId":"getUserByEmail","parameters":[{"name":"email","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/commandes/{id}/pdf":{"get":{"tags":["commande-controller"],"operationId":"generateCommandePdf","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/auth/activate-account":{"get":{"tags":["Authentication"],"operationId":"confirm","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"GrantedAuthority":{"type":"object","properties":{"authority":{"type":"string"}}},"Role":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string","enum":["ADMINISTRATEUR","PROFFESSEUR","PREPARATEUR","DIRECTEUR","USER"]},"createdDate":{"type":"string","format":"date-time"},"lastModifiedDate":{"type":"string","format":"date-time"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"firstname":{"type":"string"},"lastname":{"type":"string"},"dateOfBirth":{"type":"string","format":"date"},"email":{"type":"string"},"password":{"type":"string"},"accountLocked":{"type":"boolean"},"enabled":{"type":"boolean"},"roles":{"type":"array","items":{"$ref":"#/components/schemas/Role"}},"createdDate":{"type":"string","format":"date-time"},"lastModifiedDate":{"type":"string","format":"date-time"},"name":{"type":"string"},"username":{"type":"string"},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}},"accountNonLocked":{"type":"boolean"},"credentialsNonExpired":{"type":"boolean"},"accountNonExpired":{"type":"boolean"}}},"Armoire":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"designation":{"type":"string"},"produits":{"type":"array","items":{"$ref":"#/components/schemas/Produit"}}}},"Commande":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"designation":{"type":"string"},"date":{"type":"string","format":"date-time"},"observation":{"type":"string"},"numero":{"type":"string"},"fournisseur":{"$ref":"#/components/schemas/Fournisseur"},"user":{"$ref":"#/components/schemas/User"},"produits":{"type":"array","items":{"$ref":"#/components/schemas/ProduitCommande"}},"createdDate":{"type":"string","format":"date-time"},"lastModifiedDate":{"type":"string","format":"date-time"}}},"Fournisseur":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nom":{"type":"string"},"adresse":{"type":"string"},"email":{"type":"string"},"nmrTel":{"type":"string"},"commandes":{"type":"array","items":{"$ref":"#/components/schemas/Commande"}}}},"Preparation":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"designation":{"type":"string"},"date":{"type":"string","format":"date-time"},"quantite":{"type":"integer","format":"int32"},"produit1":{"$ref":"#/components/schemas/Produit"},"quantite1":{"type":"integer","format":"int32"},"produit2":{"$ref":"#/components/schemas/Produit"},"quantite2":{"type":"integer","format":"int32"},"preparationTPs":{"type":"array","items":{"$ref":"#/components/schemas/PreparationTP"}},"createdDate":{"type":"string","format":"date-time"},"lastModifiedDate":{"type":"string","format":"date-time"}}},"PreparationTP":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"preparation":{"$ref":"#/components/schemas/Preparation"},"quantite":{"type":"integer","format":"int32"}}},"Produit":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"designation":{"type":"string"},"reference":{"type":"string"},"type":{"type":"string","enum":["CHIMIQUE","MATERIEL","CONSOMMABLE","EQUIPEMENT","REACTIF"]},"dateExp":{"type":"string","format":"date"},"categorie":{"type":"string","enum":["CHIMIE","BIOLOGIE","PHYSIQUE","INFORMATIQUE"]},"rubrique":{"type":"string","enum":["PRODUITS_CHIMIQUES","MATERIEL_VERRERIE","INSTRUMENTS","CONSOMMABLES","SECURITE"]},"durabilite":{"type":"string","enum":["CONSOMMABLE","DURABLE","PERISSABLE","NON_PERISSABLE"]},"quantiteInitiale":{"type":"integer","format":"int32"},"fournisseur":{"$ref":"#/components/schemas/Fournisseur"},"uniteMesure":{"type":"string","enum":["ML","L","G","KG","MM","M","UNITE"]},"createdDate":{"type":"string","format":"date-time"},"lastModifiedDate":{"type":"string","format":"date-time"}}},"ProduitCommande":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"produit":{"$ref":"#/components/schemas/Produit"},"quantiteAjoutee":{"type":"integer","format":"int32"}}},"ProduitTP":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"produit":{"$ref":"#/components/schemas/Produit"},"quantite":{"type":"integer","format":"int32"}}},"SalleTp":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"numero":{"type":"string"},"armoires":{"type":"array","items":{"$ref":"#/components/schemas/Armoire"}},"createdDate":{"type":"string","format":"date-time"},"lastModifiedDate":{"type":"string","format":"date-time"}}},"Tp":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"type":{"type":"string","enum":["CHIMIE","PHYSIQUE","INFORMATIQUE","TECHNIQUE"]},"jourTp":{"type":"string"},"prof":{"$ref":"#/components/schemas/User"},"salleTp":{"$ref":"#/components/schemas/SalleTp"},"niveauScolaire":{"type":"string","enum":["NIVEAU_1","NIVEAU_2","NIVEAU_3","NIVEAU_4","NIVEAU_7","NIVEAU_8","NIVEAU_9"]},"preparations":{"type":"array","items":{"$ref":"#/components/schemas/PreparationTP"}},"produits":{"type":"array","items":{"$ref":"#/components/schemas/ProduitTP"}},"createdDate":{"type":"string","format":"date-time"},"lastModifiedDate":{"type":"string","format":"date-time"}}},"Post":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"date":{"type":"string","format":"date-time"},"content":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"lastModifiedDate":{"type":"string","format":"date-time"}}},"Labo":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"laboType":{"type":"string","enum":["TECHNIQUE","SCIENTIFIQUE","INFORMATIQUE","PHYSIQUE"]}}},"Inventaire":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"anneeScolaire":{"type":"string"},"commentaire":{"type":"string"},"date":{"type":"string","format":"date-time"},"responsable":{"type":"string"},"quantiteRestante":{"type":"integer","format":"int32"},"createdDate":{"type":"string","format":"date-time"},"lastModifiedDate":{"type":"string","format":"date-time"}}},"RegistrationRequest":{"required":["email","firstname","lastname","password"],"type":"object","properties":{"firstname":{"type":"string"},"lastname":{"type":"string"},"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"},"role":{"type":"string"}}},"AuthenticationRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}},"AuthenticationResponse":{"type":"object","properties":{"token":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"JWT auth description","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}